refactor: CrewDto 제거 및 방어적 복사를 이용한 getter 추가
- Crew, Crews의 복사 생성자 구현
